Make bWidgets compile on Windows using MSVC (including demo app)
Main issue was that MSVC doesn't support designated initializers in C++.
But also things like missing includes, issues finding OpenGL libs, etc.

Didn't actually test it yet, was using virtual box which has issues with
OpenGL.

+++ bwidgets/widgets/bwNumberSlider.h
@@ -47,7 +47,7 @@ private:
 			float min, max;
 			// Initial value before starting to edit.
 			float initial_value;
-			float precision;
+			unsigned int precision;
 		};
 		// struct { int value; ...}
 		// struct { char value; ...}
